requirements.txt
alembic==0.9.2
blueprint==3.4.2
certifi==2017.4.17
chardet==3.0.3
click==6.7
coverage==4.4.1
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gunicorn==19.7.1
iamport-rest-client==0.5.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
pluggy==0.4.0
psycopg2==2.7.1
py==1.4.34
python-dateutil==2.6.0
python-editor==1.0.3
rauth==0.7.3
requests==2.17.3
six==1.10.0
SQLAlchemy==1.1.10
urllib3==1.21.1
virtualenv==15.1.0
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
